break web composer circular dependency
