+++ apps/web/src/composerProviderState.ts
@@ -1,0 +1,62 @@
+import {
+  type ProviderKind,
+  type ProviderModelOptions,
+  type ServerProviderModel,
+} from "@t3tools/contracts";
+import {
+  isClaudeUltrathinkPrompt,
+  normalizeClaudeModelOptionsWithCapabilities,
+  normalizeCodexModelOptionsWithCapabilities,
+  resolveEffort,
+} from "@t3tools/shared/model";
+import { getProviderModelCapabilities } from "./providerModels";
+
+export type ComposerProviderStateInput = {
+  provider: ProviderKind;
+  model: string;
+  models: ReadonlyArray<ServerProviderModel>;
+  prompt: string;
+  modelOptions: ProviderModelOptions | null | undefined;
+};
+
+export type ComposerProviderState = {
+  provider: ProviderKind;
+  promptEffort: string | null;
+  modelOptionsForDispatch: ProviderModelOptions[ProviderKind] | undefined;
+  composerFrameClassName?: string;
+  composerSurfaceClassName?: string;
+  modelPickerIconClassName?: string;
+};
+
+export function getComposerProviderState(input: ComposerProviderStateInput): ComposerProviderState {
+  const { provider, model, models, prompt, modelOptions } = input;
+  const caps = getProviderModelCapabilities(models, model, provider);
+  const providerOptions = modelOptions?.[provider];
+
+  const rawEffort = providerOptions
+    ? "effort" in providerOptions
+      ? providerOptions.effort
+      : "reasoningEffort" in providerOptions
+        ? providerOptions.reasoningEffort
+        : null
+    : null;
+
+  const promptEffort = resolveEffort(caps, rawEffort) ?? null;
+  const modelOptionsForDispatch =
+    provider === "codex"
+      ? normalizeCodexModelOptionsWithCapabilities(caps, providerOptions)
+      : normalizeClaudeModelOptionsWithCapabilities(caps, providerOptions);
+  const ultrathinkActive =
+    caps.promptInjectedEffortLevels.length > 0 && isClaudeUltrathinkPrompt(prompt);
+
+  return {
+    provider,
+    promptEffort,
+    modelOptionsForDispatch,
+    ...(ultrathinkActive ? { composerFrameClassName: "ultrathink-frame" } : {}),
+    ...(ultrathinkActive
+      ? { composerSurfaceClassName: "shadow-[0_0_0_1px_rgba(255,255,255,0.04)_inset]" }
+      : {}),
+    ...(ultrathinkActive ? { modelPickerIconClassName: "ultrathink-chroma" } : {}),
+  };
+}
